Move DirectTrinoClient to a non-testing package
This originted in testing code, but it is being repurposed as a
general-purpose utility, so it should not be in a `testing` package.

The tests for this class are in the `testing/trino-tests` module in
class `io.trino.tests.TestLocalQueries`, so they are not moved along
with it.
